+++ modules/offline-validator/src/org/sbml/jsbml/validator/factory/ConstraintFactory.java
@@ -7,8 +7,8 @@ import java.util.List;
 import org.apache.log4j.Logger;
 import org.sbml.jsbml.util.StringTools;
 import org.sbml.jsbml.validator.SBMLPackage;
-import org.sbml.jsbml.validator.constraint.AnyConstraint;
-import org.sbml.jsbml.validator.constraint.ConstraintGroup;
+import org.sbml.jsbml.validator.constraints.AnyConstraint;
+import org.sbml.jsbml.validator.constraints.ConstraintGroup;
 
 public class ConstraintFactory {
     public static final int ID_EMPTY_CONSTRAINT = -1;
