+++ modules/offline-validator/src/org/sbml/jsbml/validator/factory/SpecialConstraintBuilder.java
@@ -16,9 +16,9 @@ import org.sbml.jsbml.SpeciesType;
 import org.sbml.jsbml.UnitDefinition;
 import org.sbml.jsbml.util.TreeNodeChangeListener;
 import org.sbml.jsbml.validator.ValidationContext;
-import org.sbml.jsbml.validator.constraint.AnyConstraint;
-import org.sbml.jsbml.validator.constraint.ValidationConstraint;
-import org.sbml.jsbml.validator.constraint.ValidationFunction;
+import org.sbml.jsbml.validator.constraints.AnyConstraint;
+import org.sbml.jsbml.validator.constraints.ValidationConstraint;
+import org.sbml.jsbml.validator.constraints.ValidationFunction;
 
 @SuppressWarnings("deprecation")
 public class SpecialConstraintBuilder extends AbstractConstraintBuilder {
